Move supabase file into supabase folder
